fix: coderabbit fixes
